Sound Manager added.
Balance Changes.
Bug Fixes

+++ Assets/Script/LevelSceneScripts/ArrowItem.cs
@@ -3,5 +3,5 @@
 [CreateAssetMenu]
 public class ArrowItem : SellableItem
 {
-    public int baseInterval;
+    public float baseInterval;
 }
